Add memory markers to the memory track
